Fix darwin build (privsep)
