--- bot/checkprivsep_unix.go
@@ -1,8 +1,0 @@
-// +build darwin dragonfly freebsd netbsd openbsd
-
-package bot
-
-import "log"
-
-func checkprivsep(l *log.Logger) {
-}
